% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiallelic_functions.R
\name{get_states_to_visit}
\alias{get_states_to_visit}
\title{Returns the id of states that should be visited
in a full state HMM}
\usage{
get_states_to_visit(dat, k)
}
\arguments{
\item{void}{interfunction to be documented}
}
\description{
Returns the id of states that should be visited
in a full state HMM
}
\keyword{internal}
