% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiallelic_functions.R
\name{get_states_to_visit_pair}
\alias{get_states_to_visit_pair}
\title{Get states to visit in a pair of markers}
\usage{
get_states_to_visit_pair(input.map1, input.map2, dat, pos, rf.matrix)
}
\arguments{
\item{void}{interfunction to be documented}
}
\description{
Get states to visit in a pair of markers
}
\keyword{internal}
