% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_wrappers.R
\name{hmm_map_reconstruction}
\alias{hmm_map_reconstruction}
\title{HMM estimation function}
\usage{
hmm_map_reconstruction(
  ploidy,
  n.mrk,
  n.ind,
  haplo,
  emit = NULL,
  rf_vec,
  verbose = TRUE,
  use_H0 = FALSE,
  tol = 0.001
)
}
\arguments{
\item{void}{internal function to be documented}
}
\description{
HMM estimation function
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
