% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mappoly2_to_mappoly}
\alias{mappoly2_to_mappoly}
\title{Convert biallelic, single ploidy mappoly2 data into mappoly legacy data}
\usage{
mappoly2_to_mappoly(dat)
}
\arguments{
\item{void}{internal function to be documented}
}
\description{
Convert biallelic, single ploidy mappoly2 data into mappoly legacy data
}
\examples{
p <- c(4, 4, 4, 4)
names(p) <- c("P1", "P2", "P3", "P4")
cm <- matrix(c("P1","P2","P1","P3","P2","P3","P1","P1","P2","P4","P3","P4"),
             ncol = 2, byrow = T)
D <- simulate_multiple_crosses(ploidy = p, #four parents
                               cross.mat = cm,
                               n.ind = c(20, 20, 20, 20, 20, 20), # per cross
                               n.mrk= c(10,10,10,10), # per parent
                               alleles = list(c(0:1), c(0:1), c(0:1), c(0:1)),
                               map.length = 100)
D
bipar.dat <- mappoly2_to_mappoly(D)
require(mappoly)
plot(bipar.dat[[1]])


}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
