% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappoly2.R
\docType{package}
\name{mappolymp}
\alias{mappolymp}
\title{mappolymp: Genetic mapping in interconnected families with mixed ploidy levels and multiallelic information}
\description{
Genetic mapping in interconnected families with mixed ploidy levels and multiallelic information
}
\section{Mypackage functions}{

The mypackage functions ...
}

