% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mp2_csv_to_mappoly}
\alias{mp2_csv_to_mappoly}
\title{mp2 to mp}
\usage{
mp2_csv_to_mappoly(dat)
}
\arguments{
\item{void}{internal function to be documented}
}
\description{
mp2 to mp
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
