% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiallelic_functions.R
\name{multi_2_biallelic}
\alias{multi_2_biallelic}
\title{Multiallelic to biallelic data}
\usage{
multi_2_biallelic(dat)
}
\arguments{
\item{void}{interfunction to be documented}
}
\description{
Multiallelic to biallelic data
}
\keyword{internal}
