% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiallelic_functions.R
\name{multi_2_hapmap}
\alias{multi_2_hapmap}
\title{Given multialleleic marker information, it returns
a list of phased sub-maps}
\usage{
multi_2_hapmap(p, q, g, m, mrk.names)
}
\arguments{
\item{void}{interfunction to be documented}
}
\description{
Given multialleleic marker information, it returns
a list of phased sub-maps
}
\keyword{internal}
