% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiallelic_functions.R
\name{multiallelic_rf_mat}
\alias{multiallelic_rf_mat}
\title{Returns the recombination fraction matrix for
a multiallelic set of markers}
\usage{
multiallelic_rf_mat(dat, h)
}
\arguments{
\item{void}{interfunction to be documented}
}
\description{
Returns the recombination fraction matrix for
a multiallelic set of markers
}
\keyword{internal}
