% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiallelic_functions.R
\name{plot_segreg_multiallelic}
\alias{plot_segreg_multiallelic}
\title{Plot chisq test for the whole population}
\usage{
plot_segreg_multiallelic(p.val)
}
\arguments{
\item{void}{interfunction to be documented}
}
\description{
Plot chisq test for the whole population
}
\keyword{internal}
