% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_mappoly_maps.R
\name{prepare_maps_to_merge}
\alias{prepare_maps_to_merge}
\title{Prepare maps to merge}
\usage{
prepare_maps_to_merge(map.list, data.list, parents.mat, err = 0)
}
\arguments{
\item{map.list}{a list of maps to merge. All objects must be of class
\code{mappoly.map}}

\item{data.list}{a list of objects of class \code{mappoly.data} corresponding
to the maps in argument \code{map.list}}

\item{parents.mat}{a matrix where each row contains the names of the parents
corresponding to the maps in argument \code{map.list}}

\item{err}{global genotyping error}
}
\description{
Prepare maps to merge
}
\examples{
   \donttest{
   map_ch1_BExMG <- readRDS("~/repos/current_work/rose/fullsib_maps/BExMG/map_err_ch_1.rds")
   map_ch1_SWxBE <- readRDS("~/repos/current_work/rose/fullsib_maps/SWxBE/map_err_ch_1.rds")
   map.list <- list(BExMG = map_ch1_BExMG,
                    SWxBE = map_ch1_SWxBE)
   dat_BExMG <- readRDS("~/repos/current_work/rose/data/dat_BExMG.rds")
   dat_SWxBE <- readRDS("~/repos/current_work/rose/data/dat_SWxBE.rds")
   data.list <- list(BExMG = dat_BExMG,
                     SWxBE = dat_SWxBE)
   dat <- dat_BExMG
   map_ch1_BExMG <- get_submap(map_ch1_BExMG, mrk.pos = 1:100, reestimate.rf = FALSE)
   dat <- dat_SWxBE
   map_ch1_SWxBE <- get_submap(map_ch1_SWxBE, mrk.pos = 1:100, reestimate.rf = FALSE)
   map.list <- list(BExMG = map_ch1_BExMG,
                    SWxBE = map_ch1_SWxBE)
   parents.mat <- matrix(c("BE", "MG", "SW", "BE"),
                         2, 2, byrow = TRUE,
                         dimnames = list(c("pop1", "pop2"), c("P1", "P2")))
  }

}
