% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiallelic_functions.R
\name{segreg_multiallele}
\alias{segreg_multiallele}
\title{Expected genotypic frequency}
\usage{
segreg_multiallele(x, y)
}
\arguments{
\item{void}{interfunction to be documented}
}
\description{
Expected genotypic frequency
}
\keyword{internal}
