% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiallelic_functions.R
\name{segreg_test_multiallelic}
\alias{segreg_test_multiallelic}
\title{Chisq test for the whole population}
\usage{
segreg_test_multiallelic(dat)
}
\arguments{
\item{void}{interfunction to be documented}
}
\description{
Chisq test for the whole population
}
\keyword{internal}
