% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiallelic_functions.R
\name{segreg_test_single_snp}
\alias{segreg_test_single_snp}
\title{Chisq test for a single SNP}
\usage{
segreg_test_single_snp(dat, mrk.id)
}
\arguments{
\item{void}{interfunction to be documented}
}
\description{
Chisq test for a single SNP
}
\keyword{internal}
