% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiallelic_functions.R
\name{short_to_long}
\alias{short_to_long}
\title{Short to long format}
\usage{
short_to_long(x)
}
\arguments{
\item{void}{interfunction to be documented}
}
\description{
Short to long format
}
\keyword{internal}
