% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiallelic_multiparental_simulation.R
\name{simulate_connected}
\alias{simulate_connected}
\title{Simulate haplotypes in connected populations}
\usage{
simulate_connected(
  ploidy,
  n.mrk,
  alleles,
  map.length,
  lambda = sapply(alleles, median),
  shuffle.homolog = TRUE
)
}
\arguments{
\item{void}{internal function to be documented}
}
\description{
Simulate haplotypes in connected populations
}
\examples{
h.c <- simulate_connected(ploidy = c(4,2,6),
                          n.mrk = c(10, 15, 10),
                          alleles = list(c(0:3), c(3:4), c(3:6)),
                          map.length = 10)

}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
