% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiallelic_multiparental_simulation.R
\name{simulate_cross}
\alias{simulate_cross}
\title{Simulate a biparental cross}
\usage{
simulate_cross(h1, h2, cm.map, prob1 = NULL, prob2 = NULL, seed = NULL)
}
\arguments{
\item{void}{internal function to be documented}
}
\description{
Simulate a biparental cross
}
\examples{
 ploidy1 <- 4
 ploidy2 <- 4
 n.mrk <- 7
 n.ind <- 500
 cm.map <- seq(0, 10, length.out = n.mrk)
 h1 <- simulate_multiallelic_homology_group(ploidy = ploidy1,
                                            n.mrk = n.mrk,
                                            alleles = 0:1,
                                            lambda = 1,
                                            shuffle.homolog = F)
 op <- par(mfrow = c(1,2))
 image(h1, axes = FALSE, xlab = "Markers", ylab = "homologs")
 abline(h = seq(-1/(2*ploidy1 - 2), 1 + 1/(2*ploidy1- 2),
                length.out = ploidy1+1),
        lwd = 2)
 axis(1, at = seq(0,1, length.out = n.mrk), labels = rownames(h1))
 axis(2, at = seq(0,1, length.out = ploidy1), labels = colnames(h1))

 h2 <- simulate_multiallelic_homology_group(ploidy = ploidy2,
                                            n.mrk = n.mrk,
                                            alleles = 0:1,
                                            lambda = 1,
                                            shuffle.homolog = F)
 image(h2, axes = FALSE, xlab = "Markers", ylab = "homologs")
 abline(h = seq(-1/(2*ploidy2 - 2), 1 + 1/(2*ploidy2 - 2),
                length.out = ploidy2+1),
        lwd = 2)
 axis(1, at = seq(0,1, length.out = n.mrk), labels = rownames(h2))
 axis(2, at = seq(0,1, length.out = ploidy2), labels = colnames(h2))
 par(op)
 f1 <- simulate_cross(h1, h2, cm.map)

 dose.geno <- apply(f1$offspring, c(1,3), sum)
 colnames(dose.geno) <- paste0("Ind", 1:ncol(dose.geno))
 dat <- data.frame(snp.name = paste0("M", 1:n.mrk),
                   P1 = apply(f1$ph1 , 1, sum),
                   P2 = apply(f1$ph2 , 1, sum),
                   sequence = 1,
                   sequence.pos = 1:n.mrk,
                   dose.geno)
 require(mappoly)
 dat <- table_to_mappoly(dat, 4)
 s <- make_seq_mappoly(dat, "all")
 tpt <- est_pairwise_rf(s)
 map <- est_rf_hmm_sequential(s,tpt, tol.final = 10e-5)
 print(map, detailed = TRUE)
 plot(map)
 plot_compare_haplotypes(4,
                         ph_matrix_to_list(f1$ph1[map$info$mrk.names,]),
                         ph_matrix_to_list(f1$ph2[map$info$mrk.names,]),
                         map$maps[[1]]$seq.ph$P,
                         map$maps[[1]]$seq.ph$Q)

}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
