% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiallelic_multiparental_simulation.R
\name{simulate_multiallelic_homology_group}
\alias{simulate_multiallelic_homology_group}
\title{Simulate multiallelic homology groups}
\usage{
simulate_multiallelic_homology_group(
  ploidy,
  n.mrk,
  alleles = 0:(ploidy - 1),
  lambda = median(alleles),
  shuffle.homolog = FALSE,
  seed = NULL
)
}
\arguments{
\item{void}{internal function to be documented}
}
\description{
Simulate multiallelic homology groups
}
\examples{
 ## Multiallelic
 h1 <- simulate_multiallelic_homology_group(ploidy = 4,
                                    n.mrk = 7,
                                    alleles = 3:7,
                                    lambda = 10)
 ## Biallelic with shuffling
 h2 <- simulate_multiallelic_homology_group(ploidy = 4,
                                     n.mrk = 7,
                                     alleles = 0:1,
                                     lambda = 1,
                                     shuffle.homolog = TRUE)

}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
