% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiallelic_multiparental_simulation.R
\name{simulate_multiple_crosses}
\alias{simulate_multiple_crosses}
\alias{print.mappoly2.data}
\title{Simulate multiple crosses based}
\usage{
simulate_multiple_crosses(
  ploidy,
  cross.mat,
  n.ind,
  n.mrk,
  alleles,
  map.length,
  lambda = sapply(alleles, median)
)

\method{print}{mappoly2.data}(x)
}
\arguments{
\item{void}{internal function to be documented}
}
\description{
Simulate multiple crosses based
}
\examples{
p <- c(4, 2, 6, 4)
names(p) <- c("P1", "P2", "P3", "P4")
cm <- matrix(c("P1","P2","P1","P3","P2","P3","P1","P1","P2","P4","P3","P4"), ncol = 2, byrow = T)
D <- simulate_multiple_crosses(ploidy = p, #four parents
                               cross.mat = cm,
                               n.ind = c(60, 60, 60, 60, 60, 60), #per cross
                               n.mrk= c(100,100,100,100), # per parent
                               alleles = list(c(0:3), c(4:5), c(6:11), c(12:15)),
                               map.length = 100)
D
names(D)
names(D$dat)
D$dat$`1x2`
dimnames(D$dat$`1x2`)

}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
