% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiallelic_functions.R
\name{states_for_all_markers}
\alias{states_for_all_markers}
\title{Returns the id of states that should be visited
in a full state HMM over all markers in a data
set}
\usage{
states_for_all_markers(dat)
}
\arguments{
\item{void}{interfunction to be documented}
}
\description{
Returns the id of states that should be visited
in a full state HMM over all markers in a data
set
}
\keyword{internal}
