% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states_to_visit.R
\name{states_to_visit}
\alias{states_to_visit}
\title{Generates the states the model should visit including global error rate}
\usage{
states_to_visit(input.data, err = 0, is.log = TRUE)
}
\arguments{
\item{void}{internal function to be documented}
}
\description{
Generates the states the model should visit including global error rate
}
\examples{
require(tidyverse)
p <- c(4, 2, 6, 4)
names(p) <- c("P1", "P2", "P3", "P4")
cm <- matrix(c("P1","P2","P1","P3","P2","P3","P1","P1","P2","P4","P3","P4"),
             ncol = 2, byrow = T)
D <- simulate_multiple_crosses(ploidy = p, #four parents
                               cross.mat = cm,
                               n.ind = c(20, 20, 20, 20, 20, 20), # per cross
                               n.mrk= c(10,10,10,10), # per parent
                               alleles = list(c(0:3), c(4:5), c(6:11), c(12:15)),
                               map.length = 100)
D
H <- states_to_visit(D, err = 0.05, is.log = TRUE)
names(H)
head(H$hmm.info)
mrk.sample <- names(D$map)
i1 <- H$hmm.info \%>\%
      filter(ind == "Ind_P3xP4_1", mrk \%in\% mrk.sample) \%>\%
      mutate(st.all = paste(st.p1, st.p2, sep = "-")) \%>\%
      mutate(x = 1/emit)
i1$mrk <- as_factor(i1$mrk)
p <- ggplot(i1, aes(x=st.all, y = x, fill = emit)) +
     scale_fill_distiller(palette = "RdYlBu") +
     geom_bar(stat="identity") + facet_grid(.~mrk)
p + coord_flip()

H <- states_to_visit(D, err = 0, is.log = FALSE)
head(H$hmm.info)
mrk.sample <- names(D$map)
i1 <- H$hmm.info \%>\%
      filter(ind == "Ind_P1xP2_1", mrk \%in\% mrk.sample) \%>\%
      mutate(st.all = paste(st.p1, st.p2, sep = "-"))
i1$mrk <- as_factor(i1$mrk)
p <- ggplot(i1, aes(x=st.all, y = emit, fill = emit)) +
     scale_fill_distiller(palette = "YlOrBr",direction = 1) +
     geom_bar(stat="identity") + facet_grid(.~mrk)
p + coord_flip()

}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
