% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states_to_visit.R
\name{states_to_visit_mp1}
\alias{states_to_visit_mp1}
\title{Generates the states the model should visit for mappoly legacy}
\usage{
states_to_visit_mp1(dat, ph, is.log = TRUE)
}
\arguments{
\item{void}{internal function to be documented}
}
\description{
Generates the states the model should visit for mappoly legacy
}
\examples{
require(mappoly)
n.mrk <- 2
h.temp <- sim_homologous(ploidy = 4,
                         n.mrk = n.mrk,
                         max.d = 2,
                         max.ph = 0,
                         seed = 1)

for(i in 1:n.mrk){
  h.temp$hom.allele.p[[i]] <- 1:2
  h.temp$hom.allele.q[[i]] <- 1
  h.temp$p[i] <- 2
  h.temp$q[i] <- 1
}
dat <- poly_cross_simulate(ploidy = 4,
                           rf.vec = .01,
                           n.mrk = n.mrk,
                           n.ind = 10,
                           h.temp,
                           seed = 8532,
                           draw = TRUE)
sim.map<-cumsum(c(0,rep(imf_h(.01), (n.mrk - 1))))
plot(dat)
s <- make_seq_mappoly(dat, "all")
tpt <- est_pairwise_rf(s)
tpt$pairwise$`1-2`
map <- est_rf_hmm_sequential(input.seq = s, twopt = tpt)
plot(map)
mp <- round(cumsum(mappoly::imf_h(c(0, map$maps[[1]]$seq.rf))),2)
ph <- list(ph_list_to_matrix(h.temp$hom.allele.p, dat$ploidy),
           ph_list_to_matrix(h.temp$hom.allele.q, dat$ploidy))
states.hmm <- states_to_visit_mp1(dat, ph, is.log = TRUE)
x1 <- est_map_R(states.hmm,tol = 1e-3, verbose = FALSE)
mp2 <- round(cumsum(mappoly::imf_h(c(0, x1[[2]]))), 2)
Y1 <- rbind(sim.map,mp,mp2)
rownames(Y1) <- c("simulation", "mappoly", "mappoly2")
Y1

}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
