% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiallelic_multiparental_simulation.R
\name{simulate_multiple_crosses}
\alias{simulate_multiple_crosses}
\alias{print.mappolymp.data}
\title{Simulate multiple crosses based}
\usage{
simulate_multiple_crosses(
  ploidy.vec,
  cross.mat,
  n.ind,
  n.mrk,
  alleles,
  map.length,
  lambda = sapply(alleles, median),
  shuffle.homolog = TRUE
)

\method{print}{mappolymp.data}(x)
}
\arguments{
\item{void}{internal function to be documented}
}
\description{
Simulate multiple crosses based
}
\examples{
ploidy.vec <- c(4, 2, 6, 4) #four parents
names(ploidy.vec) <- c("P1", "P2", "P3", "P4")
cross.mat <- matrix(c("P1","P2",
                      "P1","P3",
                      "P2","P3",
                      "P1","P1",
                      "P2","P4",
                      "P3","P4"), ncol = 2, byrow = T)
n.ind <- c(60, 60, 60, 60, 60, 60) #per cross
n.mrk <- c(20,15,15,25) #per parent
alleles <- list(P1 = c(1:3),
                P2 = c(4:5),
                P3 = c(6:11),
                P4 = c(12:15))
map.length <- 10 #in centimorgans
sim.cross <- simulate_multiple_crosses(ploidy.vec,
                                       cross.mat,
                                       n.ind,
                                       n.mrk,
                                       alleles,
                                       map.length)
sim.cross
names(sim.cross)
names(sim.cross$offspring)
sim.cross$offspring[1]

}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
