% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states_to_visit.R
\name{states_to_visit}
\alias{states_to_visit}
\title{Generates the states the model should visit including global error rate}
\usage{
states_to_visit(input.data)
}
\arguments{
\item{void}{internal function to be documented}
}
\description{
Generates the states the model should visit including global error rate
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
